+++ docs/source/whats_new.rst
@@ -11,7 +11,7 @@ Features
 argument. This will automatically choose the appropriate feature scale from
 the GeoAxes extent. This can also be used interactively while panning and
 zooming in a figure. :data:`cartopy.feature.NaturalEarthFeature.scale` is
-now read-only.
+now read-only. (:pull:`1102`, :pull:`983`)
 
 --------